The modal view looks much better. Need to add a prototype cell for image and text though
